Refactoring
Pulled code block stuff into separate file, moved things around for
organization
