+++ plugins/smart_quotes.py
