+++ plugins/syntax_highlighting.py
